adde axios creation boilerplate base and some other reorganization/refactoring
