+++ package.json
@@ -52,6 +52,7 @@
   },
   "dependencies": {
     "axios": "^1.11.0",
+    "axios-retry": "^4.5.0",
     "dotenv": "^16.3.1",
     "nock": "^14.0.10",
     "ts-node": "^10.9.1"
